requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
pyyaml==6.0.1
httpx==0.27.0
watchdog==4.0.0
aiosqlite==0.20.0
pydantic==2.7.1
